Change .java files

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
-        double mass = 98;
+        int mass = 98;
         double height = 1.71;
-        long index = service.calculate(height, mass);
+        int index = (int) service.calculate(height, mass);
         System.out.println(index);
     }
 }
